+++ packages/map-template/src/components/MapWrapper/MapWrapper.jsx
@@ -1,7 +1,7 @@
-import { useEffect } from "react";
+import { useEffect, useState } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import MIMap from '@mapsindoors/react-components/src/components/MIMap/MIMap';
-import { mapTypes } from "../../constants/mapTypes";
+import { mapTypes } from '../../constants/mapTypes';
 import useLiveData from '../../hooks/useLivedata';
 import mapsIndoorsInstanceState from '../../atoms/mapsIndoorsInstanceState';
 import userPositionState from '../../atoms/userPositionState';
@@ -19,10 +19,12 @@ import pitchState from '../../atoms/pitchState';
 import solutionState from '../../atoms/solutionState';
 import notificationMessageState from '../../atoms/notificationMessageState';
 import useMapBoundsDeterminer from '../../hooks/useMapBoundsDeterminer';
-import hideNonMatchesState from "../../atoms/hideNonMatchesState";
-import miTransitionLevelState from "../../atoms/miTransitionLevelState";
-import showRoadNamesState from "../../atoms/showRoadNamesState";
-import PropTypes from "prop-types";
+import hideNonMatchesState from '../../atoms/hideNonMatchesState';
+import PropTypes from 'prop-types';
+import ViewSelector from '../ViewSelector/ViewSelector';
+import LanguageSelector from '../LanguageSelector/LanguageSelector.jsx';
+import appConfigState from '../../atoms/appConfigState';
+import isNullOrUndefined from '../../helpers/isNullOrUndefined';
 
 MapWrapper.propTypes = {
     onLocationClick: PropTypes.func,
@@ -32,8 +34,10 @@ MapWrapper.propTypes = {
     onViewModeSwitchKnown: PropTypes.func.isRequired,
     resetCount: PropTypes.number.isRequired,
     mapOptions: PropTypes.object,
-    onMapOptionsChange: PropTypes.func,
-    gmMapId: PropTypes.string
+    gmMapId: PropTypes.string,
+    isWayfindingOrDirections: PropTypes.bool,
+    currentLanguage: PropTypes.string,
+    setLanguage: PropTypes.func
 };
 
 /**
@@ -54,11 +58,13 @@ let _tileStyle;
  * @param {function} onViewModeSwitchKnown - Function that is run when the view mode switch is known (if it is to be shown of not).
  * @param {number} resetCount - A counter that is incremented when the map should be reset.
  * @param {object} props.mapOptions - Options for instantiating and styling the map as well as UI elements.
- * @param {function} props.onMapOptionsChange - Function that is run when the map options are changed.
  * @param {string} props.gmMapId - Google Maps Map ID for custom styling.
+ * @param {boolean} props.isWayfindingOrDirections - Whether wayfinding or directions is active or not.
+ * @param {string} props.currentLanguage - The currently selected language code.
+ * @param {function} props.setLanguage - Function to set the selected language.
  * @returns
  */
-function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule, onMapPositionInvestigating, onViewModeSwitchKnown, resetCount, mapOptions, onMapOptionsChange, gmMapId }) {
+function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule, onMapPositionInvestigating, onViewModeSwitchKnown, resetCount, mapOptions, gmMapId, isWayfindingOrDirections, currentLanguage, setLanguage }) {
     const apiKey = useRecoilValue(apiKeyState);
     const gmApiKey = useRecoilValue(gmApiKeyState);
     const mapboxAccessToken = useRecoilValue(mapboxAccessTokenState);
@@ -75,8 +81,9 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
     const solution = useRecoilValue(solutionState);
     const [, setErrorMessage] = useRecoilState(notificationMessageState);
     const hideNonMatches = useRecoilValue(hideNonMatchesState);
-    const miTransitionLevel = useRecoilValue(miTransitionLevelState);
-    const showRoadNames = useRecoilValue(showRoadNamesState);
+    const appConfig = useRecoilValue(appConfigState);
+    const [isViewSelectorVisible, setIsViewSelectorVisible] = useState(false);
+    const [isLanguageSelectorVisible, setIsLanguageSelectorVisible] = useState(false);
 
     useLiveData(apiKey);
 
@@ -295,20 +302,26 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
     }, [tileStyle]);
 
     /**
-     * React on changes in the miTransitionLevel prop.
+     * React on changes in appConfig and sets visibility of View Selector and visibility of Language Selector.
      */
     useEffect(() => {
-        if (!isNaN(parseInt(miTransitionLevel))) {
-            onMapOptionsChange({ miTransitionLevel: miTransitionLevel })
-        }
-    }, [miTransitionLevel])
+        if (appConfig) {
+            if (isNullOrUndefined(appConfig?.appSettings?.viewSelector)) {
+                setIsViewSelectorVisible(false);
+            } else {
+                // Boolean from the App Config comes as a string. We need to return clean boolean value based on that.
+                setIsViewSelectorVisible(appConfig?.appSettings?.viewSelector.trim().toLowerCase() === 'true');
+            }
 
-    /**
-     * React on changes in the showRoadNames prop.
-     */
-    useEffect(() => {
-        onMapOptionsChange({ showRoadNames: showRoadNames })
-    }, [showRoadNames])
+
+            if (isNullOrUndefined(appConfig?.appSettings?.languageSelector)) {
+                setIsLanguageSelectorVisible(false);
+            } else {
+                // Boolean from the App Config comes as a string. We need to return clean boolean value based on that.
+                setIsLanguageSelectorVisible(appConfig?.appSettings?.languageSelector.trim().toLowerCase() === 'true');
+            }
+        }
+    }, [appConfig])
 
     return (<>
         {apiKey && <MIMap
@@ -320,6 +333,11 @@ function MapWrapper({ onLocationClick, onMapPositionKnown, useMapProviderModule,
             mapOptions={mapOptions}
             gmMapId={gmMapId}
         />}
+        {/* Pass isWayfindingOrDirections prop to ViewSelector to disable interactions while wayfinding or directions is active*/}
+        {apiKey && <>
+            <ViewSelector isViewSelectorVisible={isViewSelectorVisible} isViewSelectorDisabled={isWayfindingOrDirections} />
+            <LanguageSelector currentLanguage={currentLanguage} setLanguage={setLanguage} isVisible={isLanguageSelectorVisible} />
+        </>}
     </>)
 }