+++ packages/map-template/src/components/MapsIndoorsMap/MapsIndoorsMap.jsx
@@ -1,10 +1,8 @@
-import * as Sentry from "@sentry/react";
-import React, { useEffect, useState } from 'react';
-import { createRoutesFromChildren, matchRoutes, useLocation, useNavigationType } from 'react-dom/client';
+import { useEffect, useState } from 'react';
 import { RecoilRoot } from 'recoil';
 import MapTemplate from '../MapTemplate/MapTemplate.jsx';
-import getBooleanValue from "../../helpers/GetBooleanValue.js";
-import PropTypes from "prop-types";
+import getBooleanValue from '../../helpers/GetBooleanValue.js';
+import PropTypes from 'prop-types';
 
 MapsIndoorsMap.propTypes = {
     apiKey: PropTypes.string,
@@ -37,7 +35,9 @@ MapsIndoorsMap.propTypes = {
     showRoadNames: PropTypes.bool,
     searchExternalLocations: PropTypes.bool,
     center: PropTypes.string,
-    useAppTitle: PropTypes.bool
+    useAppTitle: PropTypes.bool,
+    showMapMarkers: PropTypes.bool,
+    mapboxMapStyle: PropTypes.string
 };
 
 /**
@@ -63,7 +63,7 @@ MapsIndoorsMap.propTypes = {
  * @param {boolean} [props.useMapProviderModule] - Set to true if the Map Template should take MapsIndoors solution modules into consideration when determining what map type to use.
  * @param {string} [props.kioskOriginLocationId] - If running the Map Template as a kiosk (upcoming feature), provide the Location ID that represents the location of the kiosk.
  * @param {number} [props.timeout] - If you want the Map Template to reset map position and UI elements to the initial state after some time of inactivity, use this to specify the number of seconds of inactivity before resetting.
- * @param {string} [props.language] - The language to show textual content in. Supported values are "en" for English, "da" for Danish, "de" for German, "fr" for French, "it" for Italian, "es" for Spanish and "nl" for Dutch. If the prop is not set, the language of the browser will be used (if it is one of the supported languages - otherwise it will default to English).
+ * @param {string} [props.language] - The language to show textual content in. Supported values are "en" for English, "da" for Danish, "de" for German, "fr" for French, "it" for Italian, "es" for Spanish, "nl" for Dutch and "zh" for Chinese. If the prop is not set, the language of the browser will be used (if it is one of the supported languages - otherwise it will default to English).
  * @param {boolean} [props.useKeyboard] - If running the Map Template as a kiosk, set this prop to true and it will prompt a keyboard.
  * @param {number} [props.miTransitionLevel] - The zoom level on which to transition from Mapbox to MapsIndoors data. Default value is 17. This feature is only available for Mapbox.
  * @param {string} [props.category] - If you want to indicate an active category on the map. The value should be the Key (Administrative ID).
@@ -74,6 +74,8 @@ MapsIndoorsMap.propTypes = {
  * @param {boolean} [props.searchExternalLocations] - If you want to perform search for external results in the Wayfinding mode. If set to true, Mapbox/Google places will be displayed depending on the Map Provider you are using. If set to false, the results returned will only be MapsIndoors results. The default is true.
  * @param {string} [props.center] - Specifies the coordinates where the map should load, represented as longitude and latitude values separated by a comma. If the specified coordinates intersect with a Venue, that Venue will be set as the current Venue.
  * @param {boolean} [props.useAppTitle] - Specifies if the Map Template should set the document title as defined in the App Config. The default value is set to false.
+ * @param {boolean} [props.showMapMarkers] - Specifies if the Map Template should show the Map Markers. The default value is set to true.
+ * @param {string} [props.mapboxMapStyle] - Specifies the Mapbox Map Style to use. The default value is set to "mapbox://styles/mapbox/standard".
  */
 function MapsIndoorsMap(props) {
 
@@ -90,8 +92,6 @@ function MapsIndoorsMap(props) {
 
         const defaultProps = {
             apiKey: 'mapspeople3d',
-            logo: 'https://app.mapsindoors.com/mapsindoors/gfx/mapspeople-logo/mapspeople-pin.svg',
-            primaryColor: '#005655', // --brand-colors-dark-pine-100 from MIDT
             useMapProviderModule: false,
             useKeyboard: false,
             searchAllVenues: false,
@@ -132,7 +132,8 @@ function MapsIndoorsMap(props) {
         const searchExternalLocationsQueryParameter = queryStringParams.get('searchExternalLocations');
         const centerQueryParameter = queryStringParams.get('center');
         const useAppTitleQueryParameter = queryStringParams.get('useAppTitle');
-
+        const showMapMarkersQueryParameter = queryStringParams.get('showMapMarkers');
+        const mapboxMapStyleQueryParameter = queryStringParams.get('mapboxMapStyle');
         // Set the initial props on the Map Template component.
 
         // For the apiKey and venue, set the venue to "AUSTINOFFICE" if the apiKey is "mapspeople3d" and no venue is provided. We want this as the default venue for the "mapspeople3d" apiKey.
@@ -146,7 +147,7 @@ function MapsIndoorsMap(props) {
             apiKey,
             venue,
             locationId: props.supportsUrlParameters && locationIdQueryParameter ? locationIdQueryParameter : props.locationId,
-            logo: props.supportsUrlParameters && logoQueryParameter ? logoQueryParameter : (props.logo || defaultProps.logo),
+            logo: props.supportsUrlParameters && logoQueryParameter ? logoQueryParameter : props.logo,
             directionsFrom: props.supportsUrlParameters && directionsFromQueryParameter ? directionsFromQueryParameter : props.directionsFrom,
             directionsTo: props.supportsUrlParameters && directionsToQueryParameter ? directionsToQueryParameter : props.directionsTo,
             tileStyle: props.supportsUrlParameters && tileStyleQueryParameter ? tileStyleQueryParameter : props.tileStyle,
@@ -155,7 +156,7 @@ function MapsIndoorsMap(props) {
             bearing: props.supportsUrlParameters && bearingQueryParameter ? bearingQueryParameter : props.bearing,
             gmApiKey: props.supportsUrlParameters && gmApiKeyQueryParameter ? gmApiKeyQueryParameter : props.gmApiKey,
             mapboxAccessToken: props.supportsUrlParameters && mapboxAccessTokenQueryParameter ? mapboxAccessTokenQueryParameter : props.mapboxAccessToken,
-            primaryColor: props.supportsUrlParameters && primaryColorQueryParameter ? '#' + primaryColorQueryParameter : (props.primaryColor || defaultProps.primaryColor),
+            primaryColor: props.supportsUrlParameters && primaryColorQueryParameter ? '#' + primaryColorQueryParameter : props.primaryColor,
             appUserRoles: props.supportsUrlParameters && appUserRolesQueryParameter ? appUserRolesQueryParameter : props.appUserRoles,
             externalIDs: props.supportsUrlParameters && externalIDsQueryParameter ? externalIDsQueryParameter : props.externalIDs,
             gmMapId: props.supportsUrlParameters && gmMapIdQueryParameter ? gmMapIdQueryParameter : props.gmMapId,
@@ -174,50 +175,18 @@ function MapsIndoorsMap(props) {
             showExternalIDs: getBooleanValue(props.supportsUrlParameters, defaultProps.showExternalIDs, props.showExternalIDs, showExternalIDsQueryParameter),
             searchExternalLocations: getBooleanValue(props.supportsUrlParameters, defaultProps.searchExternalLocations, props.searchExternalLocations, searchExternalLocationsQueryParameter),
             supportsUrlParameters: props.supportsUrlParameters,
-            useAppTitle: getBooleanValue(props.supportsUrlParameters, defaultProps.useAppTitle, props.useAppTitle, useAppTitleQueryParameter)
+            useAppTitle: getBooleanValue(props.supportsUrlParameters, defaultProps.useAppTitle, props.useAppTitle, useAppTitleQueryParameter),
+            showMapMarkers: getBooleanValue(props.supportsUrlParameters, defaultProps.showMapMarkers, props.showMapMarkers, showMapMarkersQueryParameter),
+            mapboxMapStyle: props.supportsUrlParameters && mapboxMapStyleQueryParameter ? mapboxMapStyleQueryParameter : props.mapboxMapStyle,
         });
 
     }, [props]);
 
     return (
         <RecoilRoot>
-            <Sentry.ErrorBoundary>
-                {mapTemplateProps && <MapTemplate {...mapTemplateProps}></MapTemplate>}
-            </Sentry.ErrorBoundary>
+            {mapTemplateProps && <MapTemplate {...mapTemplateProps}></MapTemplate>}
         </RecoilRoot>
     )
 }
 
-Sentry.init({
-    dsn: process.env.NODE_ENV === 'production' ? "https://0ee7fa162023d958c96db25e99c8ff6c@o351128.ingest.sentry.io/4506851619831808" : undefined,
-    // Set environment to localhost if the url includes it. Otherwise, set to production.
-    environment: window.location.hostname === 'localhost' ? 'localhost' : 'production',
-    integrations: [
-        // See docs for support of different versions of variation of react router
-        // https://docs.sentry.io/platforms/javascript/guides/react/configuration/integrations/react-router/
-        // Note: While we don't use React Router in the Map Template project, this is needed to instrument
-        // the `/` route properly. Without this, Sentry will not log anything on the main page of this app.
-        Sentry.reactRouterV6BrowserTracingIntegration({
-            useEffect: React.useEffect,
-            useLocation,
-            useNavigationType,
-            createRoutesFromChildren,
-            matchRoutes
-        }),
-        Sentry.replayIntegration()
-    ],
-
-    // Set tracesSampleRate to 1.0 to capture 100%
-    // of transactions for performance monitoring.
-    tracesSampleRate: 1.0,
-
-    // Set `tracePropagationTargets` to control for which URLs distributed tracing should be enabled
-    tracePropagationTargets: ["localhost", /^https:\/\/api\.mapsindoors\.com/],
-
-    // Disable capture Replay for sessions.
-    // Set to 100% of sessions with an error
-    replaysSessionSampleRate: 0,
-    replaysOnErrorSampleRate: 1.0,
-});
-
-export default Sentry.withProfiler(MapsIndoorsMap, { name: "MapsIndoorsMap" });
+export default MapsIndoorsMap;