+++ packages/map-template/src/i18n/de.js
@@ -78,6 +78,11 @@ const de = {
     // Opening Hours
     'Open': 'Geöffnet',
     'Closed': 'Geschlossen',
+    // View Selector
+    'Go to view': 'Ansicht wechseln',
+    'Close view selector': 'Ansichtsauswahl schließen',
+    // Language Selector
+    'Select language': 'Sprache auswählen',
 };
 
 export default de;