+++ packages/map-template/src/i18n/en.js
@@ -78,6 +78,11 @@ const en = {
     // Opening Hours
     'Open': 'Open',
     'Closed': 'Closed',
+    // View Selector
+    'Go to view': 'Go to view',
+    'Close view selector': 'Close view selector',
+    // Language Selector
+    'Select language': 'Select language',
 };
 
 export default en;