+++ packages/map-template/src/i18n/es.js
@@ -78,6 +78,11 @@ const es = {
     // Opening Hours
     'Open': 'Abierto',
     'Closed': 'Cerrado',
+    // View Selector
+    'Go to view': 'Ver ubicación',
+    'Close view selector': 'Cerrar selector de vista',
+    // Language Selector
+    'Select language': 'Seleccionar idioma',
 };
 
 export default es;